+++ lib/site_prism/timer.rb
@@ -1,15 +1,12 @@
 # frozen_string_literal: true
 
 module SitePrism
-  # [SitePrism::Timer]
   #
-  # Used to count asynchronously towards an overall desired duration or condition (Block)
+  # @api private
+  #
   class Timer
     attr_reader :wait_time
 
-    # Count towards a specified time (Supplied)
-    #
-    # @return [Proc]
     def self.run(wait_time, &block)
       new(wait_time).run(&block)
     end
@@ -19,16 +16,10 @@ module SitePrism
       @done = false
     end
 
-    # Whether the timer has completed
-    #
-    # @return [Boolean]
     def done?
       @done == true
     end
 
-    # Start the Timer and re-evaluate the block repeatedly
-    #
-    # @return [Proc]
     def run
       start
       yield self
@@ -36,9 +27,8 @@ module SitePrism
       stop
     end
 
-    # Start the Timer in a separate process
-    #
-    # Return [True]
+    private
+
     def start
       stop
       return if wait_time.zero?
@@ -50,9 +40,6 @@ module SitePrism
       end
     end
 
-    # Forcibly stop the timer, and kill any threads created by it
-    #
-    # Return [True]
     def stop
       if @thread
         @thread.kill